Move remaining router components
