+++ src/components/remotecontrol/remotecontrol.js
@@ -18,7 +18,7 @@ import './remotecontrol.scss';
 import '../../elements/emby-ratingbutton/emby-ratingbutton';
 import ServerConnections from '../ServerConnections';
 import toast from '../toast/toast';
-import { appRouter } from '../appRouter';
+import { appRouter } from '../router/appRouter';
 
 let showMuteButton = true;
 let showVolumeSlider = true;
