+++ src/index.jsx
@@ -20,7 +20,7 @@ import { appHost } from './components/apphost';
 import { getPlugins } from './scripts/settings/webSettings';
 import { pluginManager } from './components/pluginManager';
 import packageManager from './components/packageManager';
-import { appRouter, history } from './components/appRouter';
+import { appRouter, history } from './components/router/appRouter';
 import './elements/emby-button/emby-button';
 import './scripts/autoThemes';
 import './components/themeMediaPlayer';
