+++ src/plugins/comicsPlayer/plugin.js
@@ -3,7 +3,7 @@ import { Archive } from 'libarchive.js';
 import loading from '../../components/loading/loading';
 import dialogHelper from '../../components/dialogHelper/dialogHelper';
 import keyboardnavigation from '../../scripts/keyboardNavigation';
-import { appRouter } from '../../components/appRouter';
+import { appRouter } from '../../components/router/appRouter';
 import ServerConnections from '../../components/ServerConnections';
 import * as userSettings from '../../scripts/settings/userSettings';
 import { PluginType } from '../../types/plugin.ts';
