+++ src/plugins/pdfPlayer/plugin.js
@@ -3,7 +3,7 @@ import loading from '../../components/loading/loading';
 import keyboardnavigation from '../../scripts/keyboardNavigation';
 import dialogHelper from '../../components/dialogHelper/dialogHelper';
 import dom from '../../scripts/dom';
-import { appRouter } from '../../components/appRouter';
+import { appRouter } from '../../components/router/appRouter';
 import { PluginType } from '../../types/plugin.ts';
 import Events from '../../utils/events.ts';
