+++ src/plugins/youtubePlayer/plugin.js
@@ -1,5 +1,5 @@
 import browser from '../../scripts/browser';
-import { appRouter } from '../../components/appRouter';
+import { appRouter } from '../../components/router/appRouter';
 import loading from '../../components/loading/loading';
 import { setBackdropTransparency, TRANSPARENCY_LEVEL } from '../../components/backdrop/backdrop';
 import { PluginType } from '../../types/plugin.ts';
